Removing some block storage references and adding same copyright as other files

+++ tests/rc_string_array_ut/main.c
@@ -1,4 +1,5 @@
 // Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 #include <stddef.h>
 #include "testrunnerswitcher.h"
@@ -7,5 +8,5 @@ int main(void)
 {
     size_t failedTestCount = 0;
     RUN_TEST_SUITE(rc_string_array_unittests, failedTestCount);
-    return failedTestCount;
+    return (int)failedTestCount;
 }
